reworked workflow to separete rss and docx
