+++ .github/workflows/rss_and_docx_etl.yaml
@@ -7,7 +7,7 @@ env:
   SUPABASE_API_KEY: ${{ secrets.SUPABASE_API_KEY }}
 
 jobs:
-  etl-setup:
+  rss-etl:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout Repo Content
@@ -22,10 +22,26 @@ jobs:
           python -m pip install supabase
           python -m pip install python-dotenv
           python -m pip install feedparser
+      - name: Run RSS extraction script
+        run: python src/rss-etl.py  
+  docx-etl:
+    needs: rss-etl
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout Repo Content
+        uses: actions/checkout@v3
+      - name: Setup Python 
+        uses: actions/setup-python@v4
+        with: 
+          python-version: '3.11'
+      - name: Install Python Packages
+        run: |
+          python -m pip install python-docx
+          python -m pip install supabase
+          python -m pip install python-dotenv
           python -m pip install requests
       - name: Install Libreoffice for doc to docx conversion
         run:
           sudo apt install libreoffice
-      - name: Run RSS and DOCX Text extraction script
-        run: python src/web_scraper/web_scraper.py #maybe split main functions into separate files  
-        #needs: #add dependency to setup  
+      - name: Run DOCX Text extraction script
+        run: python src/rss-docx.py
